basketTotalPrice

+++ app/src/main/java/com/example/firestoreshopping/presentation/basket_page_view/BasketState.kt
@@ -6,6 +6,6 @@ data class BasketState(
     val basketList:List<Basket> = emptyList(),
     val isErorr:String="",
     val isLoading:Boolean=false,
-    val totalPrice: Double = 0.0 // Toplam fiyat için bir alan ekliyoruz
+    val totalPrice: Int = 0 // Toplam fiyat için bir alan ekliyoruz
 
 )
